# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/domain/a": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "get domain A records for given domain",
                "parameters": [
                    {
                        "minLength": 4,
                        "type": "string",
                        "description": "valid (sub)domain to query",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.DomainResult"
                            }
                        }
                    }
                }
            }
        },
        "/domain/aaaa": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "get domain AAAA (IPv6) records for given domain",
                "parameters": [
                    {
                        "minLength": 4,
                        "type": "string",
                        "description": "valid (sub)domain to query",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.DomainResult"
                            }
                        }
                    }
                }
            }
        },
        "/domain/ns": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "get NS records for given domain",
                "parameters": [
                    {
                        "minLength": 4,
                        "type": "string",
                        "description": "valid (sub)domain to query",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.DomainResult"
                            }
                        }
                    }
                }
            }
        },
        "/domain/txt": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "get TXT records for given domain",
                "parameters": [
                    {
                        "minLength": 4,
                        "type": "string",
                        "description": "valid (sub)domain to query",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK\" example([{\"registrar\":\"1.1.1.1\",\"value\":[\"exampletext\"]}])",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.TxtResult"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "structs.DomainResult": {
            "type": "object",
            "properties": {
                "record": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "21.11.25.4",
                        "1.54.23.6"
                    ]
                },
                "registrar": {
                    "type": "string",
                    "example": "1.1.1.1"
                }
            }
        },
        "structs.TxtResult": {
            "type": "object",
            "properties": {
                "registrar": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your token. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}